requirements.txt
bcrypt==5.0.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
ciscoconfparse==1.9.52
cryptography==46.0.3
deprecated==1.2.18
dnspython==2.8.0
hier-config==2.2.3
idna==3.11
invoke==2.2.1
jinja2==3.1.6
loguru==0.7.2
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
netmiko==4.6.0
ntc-templates==8.1.0
paramiko==4.0.0
passlib==1.7.4
pycparser==2.23
pygments==2.19.2
pynacl==1.6.0
pyserial==3.5
pyyaml==6.0.3
requests==2.32.5
rich==14.2.0
ruamel-yaml==0.18.16
ruamel-yaml-clib==0.2.14
scp==0.15.0
simple-term-menu==1.6.6
textfsm==2.1.0
toml==0.10.2
urllib3==2.5.0
wrapt==1.17.3
